fix: Resolved remaining `react-refresh` lint warnings
