+++ front-end-components/src/components/chart-mode/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/components/chart-mode/component";
 export * from "src/components/chart-mode/types";
