+++ front-end-components/src/components/charts/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/components/charts/horizontal-bar-chart-wrapper";
 export * from "src/components/charts/types";
