+++ front-end-components/src/components/charts/table-chart/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/components/charts/table-chart/types";
 export * from "src/components/charts/table-chart/component";
