+++ front-end-components/src/components/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/components/chart-mode";
 export * from "src/components/charts";
 export * from "src/components/chart-dimensions";
