+++ front-end-components/src/components/loading/index.tsx
@@ -1,1 +1,2 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/components/loading/component";
