+++ front-end-components/src/contexts/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/contexts/types";
 export * from "src/contexts/contexts";
