+++ front-end-components/src/services/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/services/school-api";
 export * from "src/services/types";
