+++ front-end-components/src/views/compare-your-costs/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/views/compare-your-costs/types";
 export * from "src/views/compare-your-costs/view";
