+++ front-end-components/src/views/compare-your-costs/partials/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/views/compare-your-costs/partials/total-expenditure";
 export * from "src/views/compare-your-costs/partials/expenditure-accordion";
 export * from "src/views/compare-your-costs/partials/types";
