+++ front-end-components/src/views/index.tsx
@@ -1,2 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
 export * from "src/views/compare-your-costs";
 export * from "src/views/compare-your-workforce";
